refactor: update on authentication helpers
